Refactor visual effects into separate module with improved modularity
